requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
fluidsynth==0.2
future==1.0.0
gdown==5.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.4
mido==1.3.3
numpy==1.26.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.0.0
pretty_midi==0.2.10
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.7
pytz==2025.2
requests==2.28.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
